Denormalize provider virtual machines list components
