+++ packages/forklift-console-plugin/src/modules/Providers/views/details/tabs/VirtualMachines/OVirtVirtualMachinesList.tsx
@@ -1,16 +1,37 @@
 import React from 'react';
 
-import { ResourceFieldFactory } from '@kubev2v/common';
+import { EnumToTuple, ResourceFieldFactory } from '@kubev2v/common';
 
+import { ProviderVirtualMachinesList } from './components/ProviderVirtualMachinesList';
 import { OVirtVirtualMachinesRow } from './OVirtVirtualMachinesRow';
-import {
-  ProviderVirtualMachines,
-  ProviderVirtualMachinesProps,
-  vmsFieldsMetadataFactory,
-} from './ProviderVirtualMachines';
+import { ProviderVirtualMachinesProps } from './ProviderVirtualMachines';
 
 export const oVirtVmFieldsMetadataFactory: ResourceFieldFactory = (t) => [
-  ...vmsFieldsMetadataFactory(t),
+  {
+    resourceFieldId: 'name',
+    jsonPath: '$.name',
+    label: t('Name'),
+    isVisible: true,
+    isIdentity: true, // Name is sufficient ID when Namespace is pre-selected
+    filter: {
+      type: 'freetext',
+      placeholderLabel: t('Filter by name'),
+    },
+    sortable: true,
+  },
+  {
+    resourceFieldId: 'concerns',
+    jsonPath: '$.concerns',
+    label: t('Concerns'),
+    isVisible: true,
+    sortable: true,
+    filter: {
+      type: 'enum',
+      primary: true,
+      placeholderLabel: t('Concerns'),
+      values: EnumToTuple({ Critical: 'Critical', Warning: 'Warning', Information: 'Information' }),
+    },
+  },
   {
     resourceFieldId: 'cluster',
     jsonPath: '$.vm.cluster',
@@ -74,11 +95,12 @@ export const OVirtVirtualMachinesList: React.FC<ProviderVirtualMachinesProps> = 
   loaded,
   loadError,
 }) => (
-  <ProviderVirtualMachines
+  <ProviderVirtualMachinesList
     obj={obj}
     loaded={loaded}
     loadError={loadError}
     rowMapper={OVirtVirtualMachinesRow}
     fieldsMetadataFactory={oVirtVmFieldsMetadataFactory}
+    pageId="OVirtVirtualMachinesList"
   />
 );
