+++ packages/forklift-console-plugin/src/modules/Providers/views/details/tabs/VirtualMachines/OVirtVirtualMachinesRow.tsx
@@ -1,15 +1,38 @@
 import React from 'react';
 import { TableCell } from 'src/modules/Providers/utils';
 
-import { RowProps } from '@kubev2v/common';
+import { ResourceField, RowProps } from '@kubev2v/common';
 import { OVirtVM } from '@kubev2v/types';
-import { Tr } from '@patternfly/react-table';
+import { Td, Tr } from '@patternfly/react-table';
 
-import { VMCellProps } from './components';
-import { defaultCellRenderers, renderTd, VmData } from './ProviderVirtualMachinesRow';
+import { VMCellProps, VMConcernsCellRenderer, VMNameCellRenderer } from './components';
+
+export interface VmData {
+  vm: OVirtVM;
+  name: string;
+  concerns: string;
+}
+
+const renderTd = ({ resourceData, resourceFieldId, resourceFields }: RenderTdProps) => {
+  const fieldId = resourceFieldId;
+
+  const CellRenderer = cellRenderers?.[fieldId] ?? (() => <></>);
+  return (
+    <Td key={fieldId} dataLabel={fieldId}>
+      <CellRenderer data={resourceData} fieldId={fieldId} fields={resourceFields} />
+    </Td>
+  );
+};
+
+interface RenderTdProps {
+  resourceData: VmData;
+  resourceFieldId: string;
+  resourceFields: ResourceField[];
+}
 
 const cellRenderers: Record<string, React.FC<VMCellProps>> = {
-  ...defaultCellRenderers,
+  name: VMNameCellRenderer,
+  concerns: VMConcernsCellRenderer,
   host: ({ data }) => <TableCell>{(data?.vm as OVirtVM)?.host}</TableCell>,
   cluster: ({ data }) => <TableCell>{(data?.vm as OVirtVM)?.cluster}</TableCell>,
   path: ({ data }) => <TableCell>{(data?.vm as OVirtVM)?.path}</TableCell>,
@@ -24,7 +47,7 @@ export const OVirtVirtualMachinesRow: React.FC<RowProps<VmData>> = ({
   return (
     <Tr>
       {resourceFields?.map(({ resourceFieldId }) =>
-        renderTd({ resourceData, resourceFieldId, resourceFields, cellRenderers }),
+        renderTd({ resourceData, resourceFieldId, resourceFields }),
       )}
     </Tr>
   );
