+++ packages/forklift-console-plugin/src/modules/Providers/views/details/tabs/VirtualMachines/OpenStackVirtualMachinesList.tsx
@@ -1,16 +1,37 @@
 import React from 'react';
 
-import { ResourceFieldFactory } from '@kubev2v/common';
+import { EnumToTuple, ResourceFieldFactory } from '@kubev2v/common';
 
+import { ProviderVirtualMachinesList } from './components/ProviderVirtualMachinesList';
 import { OpenStackVirtualMachinesRow } from './OpenStackVirtualMachinesRow';
-import {
-  ProviderVirtualMachines,
-  ProviderVirtualMachinesProps,
-  vmsFieldsMetadataFactory,
-} from './ProviderVirtualMachines';
+import { ProviderVirtualMachinesProps } from './ProviderVirtualMachines';
 
 export const openStackVmFieldsMetadataFactory: ResourceFieldFactory = (t) => [
-  ...vmsFieldsMetadataFactory(t),
+  {
+    resourceFieldId: 'name',
+    jsonPath: '$.name',
+    label: t('Name'),
+    isVisible: true,
+    isIdentity: true, // Name is sufficient ID when Namespace is pre-selected
+    filter: {
+      type: 'freetext',
+      placeholderLabel: t('Filter by name'),
+    },
+    sortable: true,
+  },
+  {
+    resourceFieldId: 'concerns',
+    jsonPath: '$.concerns',
+    label: t('Concerns'),
+    isVisible: true,
+    sortable: true,
+    filter: {
+      type: 'enum',
+      primary: true,
+      placeholderLabel: t('Concerns'),
+      values: EnumToTuple({ Critical: 'Critical', Warning: 'Warning', Information: 'Information' }),
+    },
+  },
   {
     resourceFieldId: 'hostID',
     jsonPath: '$.vm.hostID',
@@ -90,11 +111,12 @@ export const OpenStackVirtualMachinesList: React.FC<ProviderVirtualMachinesProps
   loaded,
   loadError,
 }) => (
-  <ProviderVirtualMachines
+  <ProviderVirtualMachinesList
     obj={obj}
     loaded={loaded}
     loadError={loadError}
     rowMapper={OpenStackVirtualMachinesRow}
     fieldsMetadataFactory={openStackVmFieldsMetadataFactory}
+    pageId="OpenStackVirtualMachinesList"
   />
 );
