+++ packages/forklift-console-plugin/src/modules/Providers/views/details/tabs/VirtualMachines/VSphereVirtualMachinesList.tsx
@@ -1,16 +1,37 @@
 import React from 'react';
 
-import { ResourceFieldFactory } from '@kubev2v/common';
+import { EnumToTuple, ResourceFieldFactory } from '@kubev2v/common';
 
-import {
-  ProviderVirtualMachines,
-  ProviderVirtualMachinesProps,
-  vmsFieldsMetadataFactory,
-} from './ProviderVirtualMachines';
+import { ProviderVirtualMachinesList } from './components/ProviderVirtualMachinesList';
+import { ProviderVirtualMachinesProps } from './ProviderVirtualMachines';
 import { VSphereVirtualMachinesRow } from './VSphereVirtualMachinesRow';
 
 export const vSphereVmFieldsMetadataFactory: ResourceFieldFactory = (t) => [
-  ...vmsFieldsMetadataFactory(t),
+  {
+    resourceFieldId: 'name',
+    jsonPath: '$.name',
+    label: t('Name'),
+    isVisible: true,
+    isIdentity: true, // Name is sufficient ID when Namespace is pre-selected
+    filter: {
+      type: 'freetext',
+      placeholderLabel: t('Filter by name'),
+    },
+    sortable: true,
+  },
+  {
+    resourceFieldId: 'concerns',
+    jsonPath: '$.concerns',
+    label: t('Concerns'),
+    isVisible: true,
+    sortable: true,
+    filter: {
+      type: 'enum',
+      primary: true,
+      placeholderLabel: t('Concerns'),
+      values: EnumToTuple({ Critical: 'Critical', Warning: 'Warning', Information: 'Information' }),
+    },
+  },
   {
     resourceFieldId: 'isTemplate',
     jsonPath: '$.vm.isTemplate',
@@ -62,11 +83,12 @@ export const VSphereVirtualMachinesList: React.FC<ProviderVirtualMachinesProps> 
   loaded,
   loadError,
 }) => (
-  <ProviderVirtualMachines
+  <ProviderVirtualMachinesList
     obj={obj}
     loaded={loaded}
     loadError={loadError}
     rowMapper={VSphereVirtualMachinesRow}
     fieldsMetadataFactory={vSphereVmFieldsMetadataFactory}
+    pageId="VSphereVirtualMachinesList"
   />
 );
