+++ packages/forklift-console-plugin/src/modules/Providers/views/details/tabs/VirtualMachines/VSphereVirtualMachinesRow.tsx
@@ -1,15 +1,38 @@
 import React from 'react';
 import { TableCell } from 'src/modules/Providers/utils';
 
-import { RowProps } from '@kubev2v/common';
+import { ResourceField, RowProps } from '@kubev2v/common';
 import { VSphereVM } from '@kubev2v/types';
-import { Tr } from '@patternfly/react-table';
+import { Td, Tr } from '@patternfly/react-table';
 
-import { VMCellProps } from './components';
-import { defaultCellRenderers, renderTd, VmData } from './ProviderVirtualMachinesRow';
+import { VMCellProps, VMConcernsCellRenderer, VMNameCellRenderer } from './components';
+
+export interface VmData {
+  vm: VSphereVM;
+  name: string;
+  concerns: string;
+}
+
+const renderTd = ({ resourceData, resourceFieldId, resourceFields }: RenderTdProps) => {
+  const fieldId = resourceFieldId;
+
+  const CellRenderer = cellRenderers?.[fieldId] ?? (() => <></>);
+  return (
+    <Td key={fieldId} dataLabel={fieldId}>
+      <CellRenderer data={resourceData} fieldId={fieldId} fields={resourceFields} />
+    </Td>
+  );
+};
+
+interface RenderTdProps {
+  resourceData: VmData;
+  resourceFieldId: string;
+  resourceFields: ResourceField[];
+}
 
 const cellRenderers: Record<string, React.FC<VMCellProps>> = {
-  ...defaultCellRenderers,
+  name: VMNameCellRenderer,
+  concerns: VMConcernsCellRenderer,
   host: ({ data }) => <TableCell>{(data?.vm as VSphereVM)?.host}</TableCell>,
   isTemplate: ({ data }) => (
     <TableCell>{Boolean((data?.vm as VSphereVM)?.isTemplate).toString()}</TableCell>
@@ -25,7 +48,7 @@ export const VSphereVirtualMachinesRow: React.FC<RowProps<VmData>> = ({
   return (
     <Tr>
       {resourceFields?.map(({ resourceFieldId }) =>
-        renderTd({ resourceData, resourceFieldId, resourceFields, cellRenderers }),
+        renderTd({ resourceData, resourceFieldId, resourceFields }),
       )}
     </Tr>
   );
