+++ packages/forklift-console-plugin/src/modules/Providers/views/details/tabs/VirtualMachines/index.ts
@@ -1,6 +1,5 @@
 // @index(['./*', /style/g], f => `export * from '${f.path}';`)
 export * from './components';
 export * from './ProviderVirtualMachines';
-export * from './ProviderVirtualMachinesRow';
 export * from './utils';
 // @endindex
